elimite group by key by fd
